Adjustments: Older SDKv2 org settings reimplemented for supporting, Changelog deprecation warning, CIDR creation function
